Adding unit tests for extent_to_json
This adds unit tests for `extent_to_json` to verify that it can create
extent JSON using both list and GeoDataFrame inputs, and that these
extent outputs are the same regardless of input type. I have also fixed
an error in the implementation of the list-based method (the elements of
the list needed to be unpacked with a `*` in order to act as positional
arguments to the `box` function.

The unit tests also include a check to ensure that a `ValueError` is
raised when we provide input that is neither a list nor a GeoDataFrame
(notice the use of the `pytest.raises` as a context manager).