Add a linestring dataset & check docstring examples
This commit adds the continental divide trail as an example dataset, so
that we have point, line, and polygon example data, along with
multichannel and single channel rasters.

Also, this now checks examples in module docstrings on Travis CI, with
instructions in the top level README for how to do it locally.

Finally, this commit updates the RGB raster data so that the single
channel red, green, and blue tifs are identical to the stacked RGB tif.