Customizing docs generation
This commit improves on the default TOC handling of API documentation.
It introduces "Package Reference" as the top-level page for API docs,
and removes the annoyong "package" and "module" suffixes that autodoc
adds by default.

In addition, it uses autodata to document data objects found in the
earthpy/__init__.py module. The docs/api/earthpy-data.rst file points to
these data objects.