https://ndownloader.figshare.com/files/21941085
I did a few things here. 1) i took the landsat data and made it public. what i'm not sure about is if that private link will still be good to use for the older scene. i'm hoping we can deprecate that one in favor of the new one which has the original data uncropped.